refactor(redux): migrate the feature editor dialog
