--- src/actions/feature-editor.js
@@ -1,12 +1,0 @@
-/**
- * @file Action factories related to the dialog that shows the feature
- * editor.
- */
-
-import { actions } from '../reducers/dialogs/feature-editor';
-
-export const {
-  closeFeatureEditorDialog,
-  setFeatureEditorDialogTab,
-  showFeatureEditorDialog,
-} = actions;
