+++ src/views/features/FeatureList.jsx
@@ -10,7 +10,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import { connect } from 'react-redux';
 
-import { showFeatureEditorDialog } from '~/actions/feature-editor';
+import { showFeatureEditorDialog } from '~/features/map-features/actions';
 import { listOf } from '~/components/helpers/lists';
 import { getNameOfFeatureType, getIconOfFeatureType } from '~/model/features';
 import { getFeaturesInOrder } from '~/selectors/ordered';
