+++ src/views/map/MapContextMenu.jsx
@@ -27,7 +27,7 @@ import Refresh from '@material-ui/icons/Refresh';
 
 import { createSelector } from '@reduxjs/toolkit';
 
-import { showFeatureEditorDialog } from '~/actions/feature-editor';
+import { showFeatureEditorDialog } from '~/features/map-features/actions';
 import { setFlatEarthCoordinateSystemOrigin } from '~/reducers/map/origin';
 
 import ContextMenu from '~/components/ContextMenu';
